fixed second issue with build

+++ aruw-mcb-project/src/aruwsrc/control/chassis/new-chassis/swerve_wheel.hpp
@@ -27,7 +27,7 @@ namespace chassis
 
 struct SwerveAzimuthConfig
 {
-    float azimuthZeroOffset;
+    int azimuthZeroOffset;
     float azimuthMotorGearing;
 };
